Different format between versions
The value for allowed-ip-addresses ExpandoColums is persisted in different format
between the versions 6.1.30 EE SP4
and 6.1.20 EE, 6.1.2 CE GA3, 6.1.1 CE GA2. While the following set of IP
rules.

159.15.129.0/24
159.15.234.143
159.15.234.0/24
159.15.234.119

is saved in database as String[] for the version 6.1.30 EE SP4, for the
other versions is saved as String being the value

159.15.129.0/24\r\n159.15.234.143\r\n159.15.234.0/24\r\n159.15.234.119\r\n127.0.0.1

This commit resolves both scenarios.